# openapi.yaml
openapi: 3.1.0
info:
  title: VigilSiddhiAI Resume Builder API
  version: 1.0.0
servers:
  - url: https://your-domain.example
paths:
...
  /api/downloads/pdf:
    get:
      summary: Download a server-generated PDF using a signed token
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PDF stream
  /api/downloads/docx:
    get:
      summary: Download a server-generated DOCX using a signed token
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: DOCX stream
...